# [bugfix] Optimize testcase queries over time period